+++ tests/test_http_server.py
@@ -12,6 +12,7 @@ from threading import Thread
 
 import pytest
 import requests
+from conftest import S3Mock
 from requests.auth import HTTPBasicAuth
 
 sys.path.append(os.path.dirname(os.path.dirname(__file__)))
@@ -26,48 +27,6 @@ def random_string(length=10):
 USER = random_string()
 PASSWORD = random_string()
 SERVER_MOCK = 'http://0.0.0.0:5000'
-
-
-class S3Mock:
-    instance = None
-    def __init__(self, *args, **kwargs):
-        if S3Mock.instance == None:
-            S3Mock.instance = self
-            self.files = {
-                'manifest': dedent("""\
-                    commands = {}
-                    modules = {}
-                    repository = {}
-                """).encode('utf-8')
-            }
-        else:
-            self.files = S3Mock.instance.files
-
-    def generate_presigned_url(self, ClientMethod, Params, ExpiresIn):
-        key = Params.get('Key')
-        if key == '/':
-            key = ''
-
-        return 'https://hb.bizmrd.ru/tarantool/%s' % key
-
-    def get_object(self, Bucket, Key):
-        return {
-            'ResponseMetadata': {'HTTPHeaders': ['content-type']}
-        }
-
-    def download_fileobj(self, Bucket, Key, Bytes):
-        Bytes.write(self.files[Key])
-
-    def upload_fileobj(self, Data, Bucket, Key):
-        logging.info('PUT %s' % Key)
-        self.files[Key] = Data.read()
-        pass
-
-    def delete_object(self, Bucket, Key):
-        logging.info('DELETE %s' % Key)
-        del self.files[Key]
-        pass
-
 
 @pytest.fixture
 def app(monkeypatch):
